Initial app code
The app is not finished. It has mostly done UI in the (editable) nib files. It has almost all "find" tool functionality done. What is missing is convenience functionality: remembering the last used options, saving and loading configs
The app is not codesigned and not notarized. In other words not ready for distribution. It requires OMC 4.0.1: https://github.com/abra-code/OMC